Add about us page and reorder pages in menu
The “About us” page is still rough (and shows the benefit of bringing
back semantic markup—the specific class names are shown to be overly
specific to product pages). The ordering of “Yard” and “Tools” is
swapped in the side/top menu, just because it seems to make more sense
this way.